added delete, update and get student by id apis

+++ src/main/java/myFirstSpringBootProject/studentRepository.java
@@ -1,5 +1,5 @@
 package myFirstSpringBootProject;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface studentRepository extends JpaRepository<student, Long> {
+public interface studentRepository extends JpaRepository<student, Integer> {
 }
